fix: fix text time to show the time remains correctly and add countdown to textTime

+++ app/src/main/java/com/gmail/devpelegrino/cronoagua/viewmodel/WaterManagementViewModel.kt
@@ -73,7 +73,7 @@ class WaterManagementViewModel(application: Application) : AndroidViewModel(appl
             _userProfile?.dailyAverage!!,
             0,
             getTime(_configuration?.wakeUpTime!!),
-            Constants.TIME_INTERVAL
+            Constants.TIME_INTERVAL.toInt()
         )
     }
